requirements.txt
ultralytics
fastapi
opencv-python
langchain-community 
langchain-core       
huggingface_hub             
faiss-cpu                   
PyPDF2                      
python-dotenv               
fastapi
